release(v3.5.1): clean up ESLint — zero lint problems
- Turn off react-refresh/only-export-components for src/components/ui/**
  (vendored shadcn primitives) and react-hooks/incompatible-library
  (non-actionable TanStack Table warning) in eslint.config.js.
- Extract SortableHeader into its own data-table/sortable-header.tsx so
  columns.tsx is a pure column-definition module.
- Fix react-hooks/exhaustive-deps in games.tsx and faceted-filter.tsx —
  wrap values built with `?? []` so dependent useMemos actually memoize.
- npm run lint now reports 0 errors, 0 warnings (was 17 + 5).
